# feat: Implement custom source formatting to improve performance