--- a/docs/docs/configuration/mapper.mdx
+++ b/docs/docs/configuration/mapper.mdx
@@ -184,7 +184,7 @@ dotnet_diagnostic.RMG020.severity = error # Unmapped source member
 
 ### Strict enum mappings
 
-To enforce strict enum mappings set `RMG037` and `RMG038` to error, see [strict enum mappings](./enum.mdx).
+To enforce strict enum mappings set `RMG037` and `RMG038` to error, see [strict enum mappings](./enum.mdx#strict-enum-mappings).
 
 ## Default Mapper configuration