--- a/src/Riok.Mapperly/Descriptors/Mappings/DerivedTypeSwitchMapping.cs
+++ b/src/Riok.Mapperly/Descriptors/Mappings/DerivedTypeSwitchMapping.cs
@@ -1,7 +1,8 @@
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
+using Riok.Mapperly.Emit.Syntax;
 using static Microsoft.CodeAnalysis.CSharp.SyntaxFactory;
-using static Riok.Mapperly.Emit.SyntaxFactoryHelper;
+using static Riok.Mapperly.Emit.Syntax.SyntaxFactoryHelper;
 
 namespace Riok.Mapperly.Descriptors.Mappings;
 
@@ -25,7 +26,7 @@ public class DerivedTypeSwitchMapping : NewInstanceMapping
     {
         // _ => throw new ArgumentException(msg, nameof(ctx.Source)),
         var sourceType = Invocation(MemberAccess(ctx.Source, GetTypeMethodName));
-        var fallbackArm = SwitchExpressionArm(
+        var fallbackArm = SwitchArm(
             DiscardPattern(),
             ThrowArgumentExpression(
                 InterpolatedString($"Cannot map {sourceType} to {TargetType.ToDisplayString()} as there is no known derived type mapping"),
@@ -38,13 +39,16 @@ public class DerivedTypeSwitchMapping : NewInstanceMapping
         var arms = _typeMappings
             .Select(x => BuildSwitchArm(typeArmVariableName, x.SourceType, x.Build(typeArmContext)))
             .Append(fallbackArm);
-        return SwitchExpression(ctx.Source).WithArms(CommaSeparatedList(arms, true));
+        return ctx.SyntaxFactory.Switch(ctx.Source, arms);
     }
 
     private SwitchExpressionArmSyntax BuildSwitchArm(string typeArmVariableName, ITypeSymbol type, ExpressionSyntax mapping)
     {
         // A x => MapToADto(x),
-        var declaration = DeclarationPattern(FullyQualifiedIdentifier(type), SingleVariableDesignation(Identifier(typeArmVariableName)));
-        return SwitchExpressionArm(declaration, mapping);
+        var declaration = DeclarationPattern(
+            FullyQualifiedIdentifier(type).AddTrailingSpace(),
+            SingleVariableDesignation(Identifier(typeArmVariableName))
+        );
+        return SwitchArm(declaration, mapping);
     }
 }