--- a/src/Riok.Mapperly/Descriptors/Mappings/SourceObjectMemberMapping.cs
+++ b/src/Riok.Mapperly/Descriptors/Mappings/SourceObjectMemberMapping.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
-using static Riok.Mapperly.Emit.SyntaxFactoryHelper;
+using static Riok.Mapperly.Emit.Syntax.SyntaxFactoryHelper;
 
 namespace Riok.Mapperly.Descriptors.Mappings;