… docs/docs/contributing/architecture.md
… docs/docs/contributing/debugging.md
… src/Riok.Mapperly/Descriptors/Mappings/LinqConstructorMapping.cs
… src/Riok.Mapperly/Descriptors/Mappings/LinqDictionaryMapping.cs
… src/Riok.Mapperly/Descriptors/Mappings/LinqEnumerableMapping.cs
… src/Riok.Mapperly/Descriptors/Mappings/MemberMappings/ConstructorParameterMapping.cs
… src/Riok.Mapperly/Descriptors/Mappings/MemberMappings/MemberAssignmentMapping.cs
… src/Riok.Mapperly/Descriptors/Mappings/MemberMappings/MemberNullAssignmentInitializerMapping.cs
… src/Riok.Mapperly/Descriptors/Mappings/MemberMappings/MemberNullDelegateAssignmentMapping.cs
… src/Riok.Mapperly/Descriptors/Mappings/MemberMappings/NullMemberMapping.cs
… src/Riok.Mapperly/Descriptors/Mappings/MemberMappings/ValueTupleConstructorParameterMapping.cs
… src/Riok.Mapperly/Descriptors/Mappings/MethodMapping.cs
… src/Riok.Mapperly/Descriptors/Mappings/NewInstanceObjectFactoryMemberMapping.cs
… src/Riok.Mapperly/Descriptors/Mappings/NewInstanceObjectMemberMapping.cs
… src/Riok.Mapperly/Descriptors/Mappings/NewInstanceObjectMemberMethodMapping.cs
… src/Riok.Mapperly/Descriptors/Mappings/NewValueTupleConstructorMapping.cs
… src/Riok.Mapperly/Descriptors/Mappings/NewValueTupleExpressionMapping.cs
… src/Riok.Mapperly/Descriptors/Mappings/NullDelegateMapping.cs
… src/Riok.Mapperly/Descriptors/Mappings/NullDelegateMethodMapping.cs
… src/Riok.Mapperly/Descriptors/Mappings/QueryableProjectionMapping.cs
… src/Riok.Mapperly/Descriptors/Mappings/StaticMethodMapping.cs
… src/Riok.Mapperly/Descriptors/Mappings/TypeMappingBuildContext.cs
… src/Riok.Mapperly/Descriptors/Mappings/UserMappings/UserDefinedExistingTargetMethodMapping.cs
… src/Riok.Mapperly/Descriptors/Mappings/UserMappings/UserDefinedNewInstanceGenericTypeMapping.cs
… src/Riok.Mapperly/Descriptors/Mappings/UserMappings/UserDefinedNewInstanceMethodMapping.cs
… src/Riok.Mapperly/Descriptors/Mappings/UserMappings/UserDefinedNewInstanceRuntimeTargetTypeMapping.cs
… src/Riok.Mapperly/Descriptors/Mappings/UserMappings/UserDefinedNewInstanceRuntimeTargetTypeParameterMapping.cs
… src/Riok.Mapperly/Descriptors/Mappings/UserMappings/UserImplementedExistingTargetMethodMapping.cs
… src/Riok.Mapperly/Descriptors/Mappings/UserMappings/UserImplementedMethodMapping.cs
… src/Riok.Mapperly/Descriptors/ObjectFactories/GenericSourceObjectFactory.cs