# Implement custom source formatting to improve performance

## Description

Implements custom source formatting during the creation of the syntax nodes.
This PR slightly changes the formatting of a few things to improve performance / readability:
* After a block no empty line is inserted
* Initializer blocks: each assignment is on its own line (even for older Roslyn versions) and has a trailing comma

<details>
  <summary>Performance current main</summary>

|       Method |        Mean |     Error |    StdDev |      Gen0 |     Gen1 |   Allocated |
|------------- |------------:|----------:|----------:|----------:|---------:|------------:|
|      Compile |    582.8 us |  11.66 us |  29.24 us |   48.8281 |  11.7188 |   302.52 KB |
| LargeCompile | 31,990.4 us | 312.57 us | 277.08 us | 2125.0000 | 875.0000 | 13270.94 KB |

</details>

<details>
  <summary>Performance with this PR</summary>

|       Method |        Mean |     Error |    StdDev |      Gen0 |     Gen1 |   Allocated |
|------------- |------------:|----------:|----------:|----------:|---------:|------------:|
|      Compile |    470.7 us |   9.38 us |  13.45 us |   47.8516 |   6.8359 |   298.45 KB |
| LargeCompile | 19,770.2 us | 187.93 us | 175.79 us | 1906.2500 | 531.2500 | 11816.03 KB |

</details>

Improvement of about 38% for our LargeCompile case and small memory improvements 🚀 
The PR is pretty big, although a lot of it are just reformatted test files and namespace moves.

## Checklist

- [x] The existing code style is followed
- [x] The commit message follows our guidelines
- [x] Performed a self-review of my code
- [x] Hard-to-understand areas of my code are commented
- [x] The documentation is updated (as applicable)
- [x] Unit tests are added/updated
- [x] Integration tests are added/updated (as applicable, especially if feature/bug depends on roslyn or framework version in use)