@latonz 2023-09-01
@TimothyMakkison thanks for you feedback 😊 
The changes on the initializer blocks are intentional and not adding an empty line after code blocks is an intentional performance wise decision. We could add it to *every* code block pretty easily. But the NormalizeWhitespace implementation only adds it if the next line is not a closing block `}`. This makes it hard as we cannot add the empty line based on the information of the node itself but would need to analyse the nodes afterwards. And adding it after each block looks strange...